refactor : 약관 타입 source of truth 통합
* TERMS_TYPES·isTermsType을 shared/model/terms-content로 이동, TERMS_CONTENT 키에서 파생
* 가입 퍼널 4개 소비처(signup 진입·ConsentStep·consent-config·약관 상세) import를 shared 모델로 통일
* 기존 signup/_shared/model/terms.ts 중복 정의 제거
